+++ src/types/index.ts
@@ -127,7 +127,6 @@ export interface MailableContract {
   mailer(name: string): this
   build(): this | Promise<this>
   render(): Promise<EmailContent>
-  renderHtml(): Promise<string | null>
 }
 
 export interface MailerContract {
@@ -136,5 +135,4 @@ export interface MailerContract {
   later(delay: number, mailable: MailableContract): Promise<any>
   bulk(mailables: MailableContract[]): Promise<any[]>
   raw(content: EmailContent): Promise<any>
-  renderHtml(mailable: MailableContract): Promise<string | null>
 }